requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
APScheduler==3.10.4
attrs==24.2.0
black==24.10.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
frozenlist==1.5.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.2.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.9.0
pyparsing==3.2.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
twilio==9.3.4
tzlocal==5.2
urllib3==2.2.3
Werkzeug==3.0.4
yarl==1.16.0
